Move from a hacky direnv cache to nix-direnv plugin
I've been using a hacked up caching speedup of direnv for a while now:
https://gist.github.com/zimbatm/77bfa0268b5a2a540c4d0d9189bb1ae2

But now I've realized that a similar approach has been released as a proper
plugin and is available in 20.03 nixpkgs, so this commit removes the old
hack and replaces it with the shiny new `nix-direnv` plugin.

More info on https://github.com/nix-community/nix-direnv/

Also, had to update to latest 20.03 nixpkgs to be able to install `nix-direnv`.